sagemaker_src/requirements.txt
torch==2.0.0
torchvision==0.15.0
Pillow==10.1.0
numpy==1.24.3
